+++ ta4j-examples/src/main/java/ta4jexamples/analysis/StrategyAnalysis.java
@@ -29,7 +29,7 @@ import org.ta4j.core.BarSeries;
 import org.ta4j.core.Strategy;
 import org.ta4j.core.TradingRecord;
 import org.ta4j.core.criteria.AverageReturnPerBarCriterion;
-import org.ta4j.core.criteria.EnterAndHoldReturnCriterion;
+import org.ta4j.core.criteria.EnterAndHoldCriterion;
 import org.ta4j.core.criteria.LinearTransactionCostCriterion;
 import org.ta4j.core.criteria.MaximumDrawdownCriterion;
 import org.ta4j.core.criteria.NumberOfBarsCriterion;
@@ -84,8 +84,8 @@ public class StrategyAnalysis {
         System.out.println("Total transaction cost (from $1000): "
                 + new LinearTransactionCostCriterion(1000, 0.005).calculate(series, tradingRecord));
         // Buy-and-hold
-        System.out
-                .println("Buy-and-hold return: " + new EnterAndHoldReturnCriterion().calculate(series, tradingRecord));
+        System.out.println("Buy-and-hold return: "
+                + new EnterAndHoldCriterion(new ReturnCriterion()).calculate(series, tradingRecord));
         // Total profit vs buy-and-hold
         System.out.println("Custom strategy return vs buy-and-hold strategy return: "
                 + new VersusEnterAndHoldCriterion(totalReturn).calculate(series, tradingRecord));